# lib/api-zod/src/generated/types/createOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderRequestDesignState } from "./createOrderRequestDesignState";

export interface CreateOrderRequest {
  /**
   * Product type (Termo, Chopera, Hoppie, …)
   * @minLength 1
   * @maxLength 120
   */
  product: string;
  /**
   * @minLength 1
   * @maxLength 120
   */
  material: string;
  /**
   * Color name + hex, e.g. "Borgoña (#6d2434)"
   * @maxLength 120
   */
  color?: string;
  /**
   * Grabado láser / Impresión UV / Plotter de corte
   * @minLength 1
   * @maxLength 120
   */
  technique: string;
  /** @maxLength 500 */
  customText?: string;
  /** @maxLength 120 */
  font?: string;
  /**
   * Only when technique is UV printing
   * @maxLength 40
   */
  textColor?: string;
  /** @maxLength 120 */
  iconName?: string;
  hasUploadedImage?: boolean;
  /** Full snapshot of the design state, enough to reproduce it */
  designState: CreateOrderRequestDesignState;
  /** Rendered preview as a base64 data URL (image/png or image/jpeg) */
  previewImage?: string;
  /** Original uploaded logo/photo as a data URL, at the highest resolution the client has. Used only to embed the artwork in the internal production file; never echoed back to the client. */
  customImage?: string;
}