# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { OrderDesignState } from "./orderDesignState";
import type { OrderStatus } from "./orderStatus";

export interface Order {
  id: string;
  orderNumber: string;
  status: OrderStatus;
  product: string;
  material: string;
  color: string | null;
  technique: string;
  customText: string | null;
  font: string | null;
  textColor: string | null;
  iconName: string | null;
  hasUploadedImage: boolean;
  designState: OrderDesignState;
  previewImagePath: string | null;
  previewImageUrl: string | null;
  /** Internal production SVG path. Only present when the request carries the internal access key; never exposed to clients. */
  productionFilePath?: string | null;
  createdAt: Date;
}